Add typings to constant
