+++ packages/grid/_modules_/grid/constants/localeTextConstants.ts
@@ -1,4 +1,6 @@
-export const DEFAULT_LOCALE_TEXT = {
+import { LocaleText } from '../models/api/localeTextApi';
+
+export const DEFAULT_LOCALE_TEXT: LocaleText = {
   toolbarDensity: 'Density',
   toolbarDensityLabel: 'Density',
   toolbarDensityCompact: 'Compact',
