+++ packages/grid/_modules_/grid/models/api/localeTextApi.ts
@@ -1,4 +1,14 @@
-import { LocaleText } from '../gridOptions';
+/**
+ * Set the types of the texts in the grid.
+ */
+export interface LocaleText {
+  // Density selector toolbar button text
+  toolbarDensity: React.ReactNode;
+  toolbarDensityLabel: string;
+  toolbarDensityCompact: string;
+  toolbarDensityStandard: string;
+  toolbarDensityComfortable: string;
+}
 
 export type LocaleTextValue = string | React.ReactNode | Function;
